(#3255) Rewrite chocolateysetup.psm1
This script has barely been touched in the last decade, in part because
it's difficult to follow. Gave it a much needed revision to remove
unnecessary clutter, make its code patterns more consistent, and be much
more sensible in how it declares and calls commands, so that we can be
much more confident that things are working how we expect them to.

This should help make it much easier to make changes to going forward,
should we ever need to.