(wip) Implement new logging API
The current implementation of the logging API is not really possible
to use in a way that does not eventually cause issues if there are
ever additional instances of the GetChocolatey class or multiple loggers
registered. The design *appears* to function as a way to register
loggers for the current GetChocolatey instance, but this is a complete
red herring.

Loggers can *only* be registered for the global application due to how
the internals of Chocolatey's logging is currently implemented. Also,
the way that the current API is coded, logging only really works for
the first couple of loggers registered; after this point, things break
down in non-obvious and non-trivial ways.

To resolve this, we can only wholesale replace the logging API. This
implementation is an improvement in functionality, but is not a full
replacement. It serves our current needs, but the logging across the
entire library needs to be reevaluated eventually.

The methods provided here are at least somewhat better at communicating
what this *does*, although we do acknowledge that it is not ideal to
have these presented as builder methods on GetChocolatey rather than
a static method more obviously not attached to a given individual
instance.